# Use independent Melspec and Encoder models for iPhones 