### Why is this change needed?
Compile time Metrics on iPhone 16 Pro Max:

Before, 20,000ms (5mins+) to compile encoder and processor

```
Updating content for activity 34DE0CCF-64A6-455F-BE25-20AFCB9CFDC4
Updating content for activity 34DE0CCF-64A6-455F-BE25-20AFCB9CFDC4
Loaded model: MelEncoder.mlmodelc
Compiled ASR model MelEncoder.mlmodelc in 20018.84 ms
Updating content for activity 34DE0CCF-64A6-455F-BE25-20AFCB9CFDC4
Loaded model: Decoder.mlmodelc
Compiled ASR model Decoder.mlmodelc in 116.61 ms
Loaded model: JointDecision.mlmodelc
Compiled ASR model JointDecision.mlmodelc in 122.83 ms
Successfully downloaded ASR models
```

After: 
```
Loaded model: Encoder.mlmodelc
Compiled ASR model Encoder.mlmodelc in 3361.23 ms
Found parakeet-tdt-0.6b-v3-coreml locally, no download needed
Updating content for activity 18AB8F87-3186-4EE1-BE7D-C2F5893E29E6
Loaded model: Decoder.mlmodelc
Compiled ASR model Decoder.mlmodelc in 88.49 ms
Found parakeet-tdt-0.6b-v3-coreml locally, no download needed
Loaded model: JointDecision.mlmodelc
Compiled ASR model JointDecision.mlmodelc in 48.46 ms
Successfully downloaded ASR models
Loading ASR models from: /var/mobile/Containers/Data/Application/94B9177E-C7FC-465C-82E9-7BF21F0F971A/Library/Application Support/FluidAudio/Models/parakeet-tdt-0.6b-v3-coreml
Found parakeet-tdt-0.6b-v3-coreml locally, no download needed
Loaded model: Preprocessor.mlmodelc
Compiled ASR model Preprocessor.mlmodelc in 9.15 ms
Loaded Preprocessor.mlmodelc with compute units: MLComputeUnits(rawValue: 2)
Found parakeet-tdt-0.6b-v3-coreml locally, no download needed
Loaded model: Encoder.mlmodelc
Compiled ASR model Encoder.mlmodelc in 162.05 ms
Loaded Encoder.mlmodelc with compute units: MLComputeUnits(rawValue: 1)
Found parakeet-tdt-0.6b-v3-coreml locally, no download needed
Loaded model: Decoder.mlmodelc
Compiled ASR model Decoder.mlmodelc in 8.11 ms
Loaded Decoder.mlmodelc with compute units: MLComputeUnits(rawValue: 1)
Found parakeet-tdt-0.6b-v3-coreml locally, no download needed
Loaded model: JointDecision.mlmodelc
Compiled ASR model JointDecision.mlmodelc in 7.97 ms
Loaded JointDecision.mlmodelc with compute units: MLComputeUnits(rawValue: 1)
Loaded vocabulary with 8192 tokens from /var/mobile/Containers/Data/Application/94B9177E-C7FC-465C-82E9-7BF21F0F971A/Library/Application Support/FluidAudio/Models/parakeet-tdt-0.6b-v3-coreml/parakeet_v3_vocab.json
Successfully loaded all ASR models with optimized compute units
Initializing AsrManager with provided models
Token duration optimization model loaded successfully

```